Took care of a bunch of smaller issues and added documentation, and fixed a big one.
I was really annoyed by the fact that the current implementation would
wrap the route functions in production as well, leading to some
shortcut artefacts being visible in stack traces and stuff. So I
added the skip to the `__init__` instead, which sadly meant that
I had to say goodbye to the cute `__call__`, since those can't
be patched on an instance level.